+++ NEWS.md
@@ -6,6 +6,7 @@
 * implemented `convert_wideToLong()` function to convert trackdata tibble objects to their long form representation (useful for spectral analysis)
 * `query()` sub-function now sets `perl = TRUE` arguments in `regexp` call when using `sessionPattern` and `bundlePattern` (useful for e.g. for negative look-around `bundlePattern = "^((?!msajc003).)*$"`)
 * implemented `export_seglistToTxtCollection()` 
+* normalize length now also using `list()` with `do.call(rbind, res_list)`
 
 ## bug fixes